Project1-WeatherRadar/merra-api/requirements.txt
matplotlib~=3.5.1
numpy
pandas
scipy
setuptools==60.6.0
pytz~=2021.3
nexradaws~=1.1
netCDF4
fastapi~=0.73.0
uvicorn~=0.17.1
Pillow~=9.0.0
arm-pyart
pooch
